Use go-metrics in status monitor
Use package 'github.com/rcrowley/go-metrics' to manage metrics recording in the
node status monitor. Current functionality is exactly the same as it was before;
however, this new factoring will make it much easier to add more advanced
statistics such as moving averages and histograms.

This refactoring also moves the responsibility of naming metrics to the Monitor,
when it was previously in the recorder; this should also make managing larger
sets of statistics more straightforward.

Finally, this commit removes the `rangeDataAccumulator` interface completely;
having moved to a table-based system, we clearly are not intending to gather
statistics based on accounting prefixes, which was the original intention of
this system. This is a significant simplification; a subsequent commit should
completely remove the "beginScanRanges" and "endScanRanges" events from the
feed.

+++ storage/feed.go
@@ -103,9 +103,9 @@ type ReplicationStatusEvent struct {
 	// periodically polling the ranges of each store.
 	// TODO(mrtracy): See if this information could be computed incrementally
 	// from other events.
-	LeaderRangeCount     int32
-	ReplicatedRangeCount int32
-	AvailableRangeCount  int32
+	LeaderRangeCount     int64
+	ReplicatedRangeCount int64
+	AvailableRangeCount  int64
 }
 
 // BeginScanRangesEvent occurs when the store is about to scan over all ranges.
@@ -187,7 +187,7 @@ func (sef StoreEventFeed) storeStatus(desc *roachpb.StoreDescriptor) {
 }
 
 // replicationStatus publishes a ReplicationStatusEvent to this feed.
-func (sef StoreEventFeed) replicationStatus(leaders, replicated, available int32) {
+func (sef StoreEventFeed) replicationStatus(leaders, replicated, available int64) {
 	sef.f.Publish(&ReplicationStatusEvent{
 		StoreID:              sef.id,
 		LeaderRangeCount:     leaders,
